Feature/wip after pull

+++ backend/src/main/java/com/inflowia/medicflow/dto/consulta/ConsultaMinDTO.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -19,7 +20,7 @@ public class ConsultaMinDTO {
     private LocalDateTime dataHora;
     private TipoConsulta tipo;
     private StatusConsulta status;
-    private Double valorConsulta;
+    private BigDecimal valorConsulta;
     private Boolean pago;
     private Long pacienteId;
     private Long medicoId;
